requirements.txt
asgiref==3.7.2
Django==3.2.20
django-appconf==1.0.5
django-select2==7.11.1
Pillow==9.5.0
pytz==2023.3
reportlab==4.0.4
sqlparse==0.4.4
typing_extensions==4.7.1
